+++ src/test/java/mil/nga/geopackage/ReadmeTest.java
@@ -126,8 +126,7 @@ public class ReadmeTest extends CreateGeoPackageTestCase {
 		FeatureDao featureDao = geoPackage.getFeatureDao(featureTable);
 		FeatureResultSet featureResultSet = featureDao.queryForAll();
 		try {
-			while (featureResultSet.moveToNext()) {
-				FeatureRow featureRow = featureResultSet.getRow();
+			for (FeatureRow featureRow : featureResultSet) {
 				GeoPackageGeometryData geometryData = featureRow.getGeometry();
 				if (geometryData != null && !geometryData.isEmpty()) {
 					Geometry geometry = geometryData.getGeometry();
@@ -143,8 +142,7 @@ public class ReadmeTest extends CreateGeoPackageTestCase {
 		TileDao tileDao = geoPackage.getTileDao(tileTable);
 		TileResultSet tileResultSet = tileDao.queryForAll();
 		try {
-			while (tileResultSet.moveToNext()) {
-				TileRow tileRow = tileResultSet.getRow();
+			for (TileRow tileRow : tileResultSet) {
 				byte[] tileBytes = tileRow.getTileData();
 				BufferedImage tileImage = tileRow.getTileDataImage();
 				// ...