+++ src/test/java/mil/nga/geopackage/features/index/FeatureIndexManagerImportTest.java
@@ -47,6 +47,19 @@ public class FeatureIndexManagerImportTest extends ImportGeoPackageTestCase {
 	}
 
 	/**
+	 * Test index pagination
+	 *
+	 * @throws SQLException
+	 *             upon error
+	 */
+	@Test
+	public void testIndexPagination() throws SQLException {
+
+		FeatureIndexManagerUtils.testIndexPagination(geoPackage);
+
+	}
+
+	/**
 	 * Test large index
 	 *
 	 * @throws SQLException