+++ src/test/java/mil/nga/geopackage/tiles/user/TileUtils.java
@@ -107,9 +107,8 @@ public class TileUtils {
 				TileResultSet cursor = dao.queryForAll();
 				int count = cursor.getCount();
 				int manualCount = 0;
-				while (cursor.moveToNext()) {
+				for(TileRow tileRow: cursor) {
 
-					TileRow tileRow = cursor.getRow();
 					validateTileRow(dao, columns, tileRow);
 
 					manualCount++;