+++ CHANGELOG.md
@@ -11,7 +11,8 @@ Adheres to [Semantic Versioning](http://semver.org/).
 * tiff version 3.0.0
 * TileGenerator individual zoom level specifications
 * TileCreator dynamic separated tile width and height determination adjustment
-* FeatureIndexManager paginated query support
+* Iterable row results
+* Improved pagination query support
 
 ## [6.1.2](https://github.com/ngageoint/geopackage-java/releases/tag/6.1.2) (12-17-2021)